# src/main/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DiffEntryView": {
        "description": "One changed path between the served baseline and the pinned commit",
        "properties": {
          "binary": {
            "description": "True when either side is binary; such an entry carries no diff text",
            "type": "boolean"
          },
          "diff": {
            "description": "Unified text diff, or null for a binary entry or when no baseline is served",
            "type": "string"
          },
          "path": {
            "description": "Path within the snapshot",
            "type": "string"
          },
          "truncated": {
            "description": "True when the diff text was cut at the per-file size limit",
            "type": "boolean"
          },
          "type": {
            "description": "How the path changed relative to the served baseline",
            "enum": [
              "added",
              "modified",
              "removed"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "FileContent": {
        "description": "One blob of the pinned commit, as text for rendering only",
        "properties": {
          "binary": {
            "description": "True for a blob detected as binary; such a blob carries no text",
            "type": "boolean"
          },
          "path": {
            "description": "Path within the snapshot",
            "type": "string"
          },
          "size": {
            "description": "Full blob size in bytes",
            "format": "int64",
            "type": "integer"
          },
          "snapshotId": {
            "description": "Snapshot id",
            "format": "int64",
            "type": "integer"
          },
          "text": {
            "description": "Blob content as UTF-8 text, or null for a binary blob",
            "type": "string"
          },
          "truncated": {
            "description": "True when the text was cut at the per-file size limit",
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "FileTree": {
        "description": "The file tree of exactly the commit the snapshot pins",
        "properties": {
          "entries": {
            "description": "Paths in the pinned commit's tree",
            "items": {
              "$ref": "#/components/schemas/TreeEntry"
            },
            "type": "array"
          },
          "sha": {
            "description": "Pinned commit SHA",
            "type": "string"
          },
          "snapshotId": {
            "description": "Snapshot id",
            "format": "int64",
            "type": "integer"
          },
          "truncated": {
            "description": "True when the listing was cut at the tree-size limit",
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "HeldContentQuery": {
        "description": "One marketplace-and-commit pair a client holds",
        "properties": {
          "marketplace": {
            "description": "Marketplace name, as it appears in the fetch URL",
            "type": "string"
          },
          "sha": {
            "description": "The commit the client holds",
            "type": "string"
          }
        },
        "type": "object"
      },
      "HeldContentReport": {
        "description": "One state per requested pair, in the order asked",
        "properties": {
          "results": {
            "description": "The answers",
            "items": {
              "$ref": "#/components/schemas/HeldContentAnswer"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "HeldContentRequest": {
        "description": "The content a client holds and wants a statement about",
        "properties": {
          "holdings": {
            "description": "Marketplace-and-commit pairs the client holds, at most 256",
            "items": {
              "$ref": "#/components/schemas/HeldContentQuery"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "Run": {
        "description": "One execution of the vetting chain against a snapshot",
        "properties": {
          "chain": {
            "description": "Identity of the chain that produced the run: vetter@version in chain order",
            "example": "prompt-injection@1,secret-scan@1",
            "type": "string"
          },
          "finishedAt": {
            "description": "When the run finished, or null if it never did",
            "format": "date-time",
            "type": "string"
          },
          "outcome": {
            "description": "Fail-closed aggregate of the run's verdicts",
            "enum": [
              "clear",
              "clear_with_waivers",
              "blocked"
            ],
            "type": "string"
          },
          "runId": {
            "description": "Chain run id",
            "format": "int64",
            "type": "integer"
          },
          "snapshotId": {
            "description": "Snapshot the run vetted",
            "format": "int64",
            "type": "integer"
          },
          "startedAt": {
            "description": "When the run started",
            "format": "date-time",
            "type": "string"
          },
          "trigger": {
            "description": "What caused the run",
            "enum": [
              "ingestion",
              "revet-scheduled",
              "revet-manual"
            ],
            "type": "string"
          },
          "verdicts": {
            "description": "The run's verdicts, in chain order",
            "items": {
              "$ref": "#/components/schemas/VerdictView"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "Snapshot": {
        "description": "An immutable, SHA-identified snapshot of an upstream marketplace",
        "properties": {
          "createdAt": {
            "description": "Ingestion time",
            "format": "date-time",
            "type": "string"
          },
          "decidedAt": {
            "description": "Decision time, or null while held",
            "format": "date-time",
            "type": "string"
          },
          "decidedBy": {
            "description": "Reviewer who decided, or null while held",
            "type": "string"
          },
          "deletedAt": {
            "description": "When the snapshot was soft-deleted, or null when it is live",
            "format": "date-time",
            "type": "string"
          },
          "deletedReason": {
            "description": "Retention criterion or administrative reason for the deletion",
            "type": "string"
          },
          "id": {
            "description": "Snapshot id",
            "format": "int64",
            "type": "integer"
          },
          "ingestedBy": {
            "description": "Identity that triggered the ingestion: a principal for an on-demand ingest or a push, 'scheduler' or 'webhook' for an automated trigger, null for a snapshot ingested before the actor was recorded",
            "type": "string"
          },
          "marketplaceId": {
            "description": "Owning marketplace id",
            "format": "int64",
            "type": "integer"
          },
          "purgeAfter": {
            "description": "End of the restore window; after it compaction removes the snapshot permanently",
            "format": "date-time",
            "type": "string"
          },
          "revokedAt": {
            "description": "When re-vetting revoked the snapshot, or null",
            "format": "date-time",
            "type": "string"
          },
          "revokedBy": {
            "description": "Identity that revoked it, or null",
            "type": "string"
          },
          "revokedKind": {
            "description": "Whether a re-vetting run or an administrator revoked it, or null if it never was. The two lift differently: a re-vetting revocation lifts when its finding is cleared, an administrative one only when a second administrator reverses it",
            "enum": [
              "revet",
              "administrative"
            ],
            "type": "string"
          },
          "sha": {
            "description": "Commit SHA the snapshot is pinned to and serves: the upstream commit, or the synthesised composite when external plugin sources were resolved",
            "type": "string"
          },
          "state": {
            "description": "held, approved, rejected, or revoked (retroactively quarantined by re-vetting)",
            "enum": [
              "held",
              "approved",
              "rejected",
              "revoked"
            ],
            "type": "string"
          },
          "upstreamSha": {
            "description": "Commit SHA ingested from upstream; equal to sha unless a composite was synthesised",
            "type": "string"
          },
          "violation": {
            "description": "Policy or vetting violation that rejected or revoked the snapshot, or null",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SnapshotDiff": {
        "description": "The snapshot's delta against the marketplace's currently served commit",
        "properties": {
          "baselineSha": {
            "description": "The served commit the diff is against, or null when nothing is served",
            "type": "string"
          },
          "entries": {
            "description": "Changed paths; with no baseline, every path of the snapshot, all added",
            "items": {
              "$ref": "#/components/schemas/DiffEntryView"
            },
            "type": "array"
          },
          "sha": {
            "description": "Pinned commit SHA",
            "type": "string"
          },
          "snapshotId": {
            "description": "Snapshot id",
            "format": "int64",
            "type": "integer"
          },
          "truncated": {
            "description": "True when the entry list was cut at the diff-size limit",
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "Suppression": {
        "description": "A finding an active waiver is currently suppressing",
        "properties": {
          "approvedBy": {
            "description": "Identity that accepted the risk",
            "type": "string"
          },
          "expiresAt": {
            "description": "When the acceptance lapses",
            "format": "date-time",
            "type": "string"
          },
          "location": {
            "description": "Where the finding was located",
            "type": "string"
          },
          "ruleId": {
            "description": "Finding rule identifier",
            "type": "string"
          },
          "vetter": {
            "description": "Vetter whose verdict carried the finding",
            "type": "string"
          },
          "waiverId": {
            "description": "Waiver suppressing it",
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "TreeEntry": {
        "description": "One path in the pinned commit's tree",
        "properties": {
          "path": {
            "description": "Path within the snapshot",
            "type": "string"
          },
          "size": {
            "description": "Blob size in bytes",
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "UncoveredFinding": {
        "description": "A blocking finding that no active waiver covers",
        "properties": {
          "location": {
            "description": "Where the finding was located",
            "type": "string"
          },
          "message": {
            "description": "Reviewer-facing explanation",
            "type": "string"
          },
          "ruleId": {
            "description": "Finding rule identifier",
            "type": "string"
          },
          "severity": {
            "description": "How much it matters",
            "enum": [
              "info",
              "low",
              "medium",
              "high",
              "critical"
            ],
            "type": "string"
          },
          "vetter": {
            "description": "Vetter whose verdict carried the finding",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "VetterView": {
        "description": "A vetter configured in the vetting chain",
        "properties": {
          "description": {
            "description": "What the vetter looks for, and what it cannot see",
            "type": "string"
          },
          "external": {
            "description": "Whether the verdict is delegated to an operator-configured external service, so that a verdict the gateway did not reach itself never reads as one that it did",
            "type": "boolean"
          },
          "name": {
            "description": "Stable vetter name",
            "type": "string"
          },
          "order": {
            "description": "Position in the chain",
            "format": "int32",
            "type": "integer"
          },
          "version": {
            "description": "Identity of the rule set the vetter currently carries (GW_VETTING_0012)",
            "type": "string"
          }
        },
        "type": "object"
      },
      "VettingOverrideRecord": {
        "description": "An administrator's override of a blocked vetting outcome on a snapshot",
        "properties": {
          "blockingVetters": {
            "description": "The vetters that were blocking",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "description": "Override id",
            "format": "int64",
            "type": "integer"
          },
          "overriddenAt": {
            "description": "When the override was recorded",
            "format": "date-time",
            "type": "string"
          },
          "overriddenBy": {
            "description": "Identity of the administrator who overrode the block",
            "type": "string"
          },
          "reason": {
            "description": "The administrator's stated reason",
            "type": "string"
          },
          "snapshotId": {
            "description": "The snapshot approved over the block",
            "format": "int64",
            "type": "integer"
          },
          "uncoveredFindings": {
            "description": "Summary of the blocking findings no waiver covered",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "VettingView": {
        "description": "A snapshot's latest vetting chain run, the waivers over it, and the chain that produced it",
        "properties": {
          "outcome": {
            "description": "The effective outcome, which is what gates approval: the run's verdicts with every waived finding removed. `clear_with_waivers` means nothing objects any more only because an active waiver is suppressing a finding. A snapshot with no run is blocked.",
            "enum": [
              "clear",
              "clear_with_waivers",
              "blocked"
            ],
            "type": "string"
          },
          "override": {
            "$ref": "#/components/schemas/VettingOverrideRecord",
            "description": "Present when an administrator approved this snapshot over a blocked vetting outcome (GW_VETTING_0028); its presence is what surfaces the override so it is never indistinguishable from a clean approval. Null otherwise."
          },
          "recordedOutcome": {
            "description": "What the vetters themselves concluded, before any waiver was applied",
            "enum": [
              "clear",
              "blocked"
            ],
            "type": "string"
          },
          "run": {
            "$ref": "#/components/schemas/Run",
            "description": "The latest run with its verdicts and findings, or null if the chain never ran"
          },
          "snapshotId": {
            "description": "Snapshot id",
            "format": "int64",
            "type": "integer"
          },
          "suppressed": {
            "description": "Findings an active waiver is currently suppressing, keyed by vetter, rule and location",
            "items": {
              "$ref": "#/components/schemas/Suppression"
            },
            "type": "array"
          },
          "uncovered": {
            "description": "Blocking findings that no active waiver covers; the waivers approval still needs",
            "items": {
              "$ref": "#/components/schemas/UncoveredFinding"
            },
            "type": "array"
          },
          "vetters": {
            "description": "The vetters configured in the chain, in the order they run",
            "items": {
              "$ref": "#/components/schemas/VetterView"
            },
            "type": "array"
          },
          "waivers": {
            "description": "Waivers of this snapshot's marketplace whose rule appears in this run, active and lapsed alike, so an expired acceptance stays visible",
            "items": {
              "$ref": "#/components/schemas/WaiverView"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "WaiverView": {
        "description": "A waiver as returned by the API, with its activity resolved as of now",
        "properties": {
          "active": {
            "description": "Whether the waiver suppresses anything right now",
            "type": "boolean"
          },
          "approvedBy": {
            "description": "Identity that accepted the risk",
            "type": "string"
          },
          "createdAt": {
            "description": "When the waiver was created",
            "format": "date-time",
            "type": "string"
          },
          "expiresAt": {
            "description": "When the acceptance lapses; never null",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Waiver id",
            "format": "int64",
            "type": "integer"
          },
          "justification": {
            "description": "Why the risk is accepted",
            "type": "string"
          },
          "marketplace": {
            "description": "Marketplace the waiver belongs to",
            "type": "string"
          },
          "revokedAt": {
            "description": "When the waiver was revoked, or null",
            "format": "date-time",
            "type": "string"
          },
          "revokedBy": {
            "description": "Identity that revoked it, or null",
            "type": "string"
          },
          "ruleId": {
            "description": "Finding rule identifier being accepted",
            "example": "aws-access-key-id",
            "type": "string"
          },
          "scope": {
            "description": "How the scope value is matched",
            "enum": [
              "snapshot",
              "path"
            ],
            "type": "string"
          },
          "scopeValue": {
            "description": "A commit SHA for snapshot scope, a repository-relative path for path scope",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
...
  "info": {
    "contact": {
      "name": "jimisola",
      "url": "https://github.com/skillsgateway/skillsgateway"
    },
...
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Skills Gateway",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/snapshots/{id}": {
      "delete": {
        "description": "Marks the snapshot deleted and restorable until the end of the marketplace's restore window; its vetting state is unchanged. Approved snapshots are served by the git facade and can never be deleted.",
        "operationId": "softDelete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            },
            "description": "Snapshot marked deleted"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            },
            "description": "Snapshot not found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            },
            "description": "Snapshot is approved, or already deleted"
          }
        },
        "summary": "Soft-delete a snapshot",
        "tags": [
          "Retention"
        ]
      }
    },
...
    "/api/v1/snapshots/{id}/diff": {
      "get": {
        "description": "Added, modified and removed paths between the pinned commit and the marketplace's currently served commit (the published repository's served tip), with a unified text diff per non-binary entry under the same size caps as file reads. When the marketplace serves nothing the baseline is null and every path is reported as added. Privileged.",
        "operationId": "diff",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotDiff"
                }
              }
            },
            "description": "The delta a reviewer decides"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotDiff"
                }
              }
            },
            "description": "The session holds no applicable role"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotDiff"
                }
              }
            },
            "description": "Snapshot not found"
          }
        },
        "summary": "Diff against the currently served commit",
        "tags": [
          "Snapshot preview"
        ]
      }
    },
...
    "/api/v1/snapshots/{id}/file": {
      "get": {
        "description": "One blob addressed strictly within the pinned commit's tree — a path the tree does not contain, traversal shapes included, is not found. Text is returned for rendering only, cut at 128 KiB with an explicit truncation marker; a blob detected as binary returns metadata without text. Privileged.",
        "operationId": "file",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FileContent"
                }
              }
            },
            "description": "Blob metadata, and its text unless binary"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FileContent"
                }
              }
            },
            "description": "The session holds no applicable role"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FileContent"
                }
              }
            },
            "description": "Snapshot not found, or the path is not in the pinned tree"
          }
        },
        "summary": "One file of the pinned commit",
        "tags": [
          "Snapshot preview"
        ]
      }
    },
    "/api/v1/snapshots/{id}/files": {
      "get": {
        "description": "Every path in exactly the commit the snapshot pins, resolved through the quarantine repository's object store, capped at 2000 entries with an explicit marker when cut. Privileged: admin or an approver of the snapshot's marketplace.",
        "operationId": "files",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FileTree"
                }
              }
            },
            "description": "Paths and sizes of the pinned commit's tree"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FileTree"
                }
              }
            },
            "description": "The session holds no applicable role"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FileTree"
                }
              }
            },
            "description": "Snapshot not found"
          }
        },
        "summary": "File tree of the pinned commit",
        "tags": [
          "Snapshot preview"
        ]
      }
    },
...
    "/api/v1/snapshots/{id}/vetting": {
      "get": {
        "description": "The snapshot's latest vetting chain run: each vetter's verdict in chain order, the findings behind it, the waivers currently suppressing any of them, and the fail-closed effective aggregate that gates approval. A snapshot the chain has never run against reports a blocked outcome and no run.",
        "operationId": "vetting",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VettingView"
                }
              }
            },
            "description": "The latest chain run and the configured chain"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VettingView"
                }
              }
            },
            "description": "Snapshot not found"
          }
        },
        "summary": "Snapshot vetting verdicts",
        "tags": [
          "Vetting"
        ]
      }
    },
...
    "/status/v1/snapshots": {
      "post": {
        "description": "Answers one state per marketplace-and-commit pair, in the order asked. A pair the gateway makes no positive statement about is answered 'unknown', which is never a pass: a client that holds such content should treat it as withdrawn. Authenticates with the same credentials the git facade accepts, and answers only about marketplaces that credential could fetch.",
        "operationId": "status",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeldContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HeldContentReport"
                }
              }
            },
            "description": "One state per requested pair, in the order asked"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HeldContentReport"
                }
              }
            },
            "description": "Malformed request, or more pairs than one request may carry"
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HeldContentReport"
                }
              }
            },
            "description": "Unauthenticated"
          }
        },
        "summary": "Ask whether content a client already holds is still approved",
        "tags": [
          "Held content"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ],
...
}